logic refactor
