+++ .idea/.idea.Tanks_patterns/.idea/codeStyles/codeStyleConfig.xml
@@ -1,0 +1,5 @@
+﻿<component name="ProjectCodeStyleConfiguration">
+  <state>
+    <option name="PREFERRED_PROJECT_CODE_STYLE" value="Default" />
+  </state>
+</component>
