+++ .idea/.idea.Tanks_patterns/.idea/contentModel.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ContentModelStore">
-    <e p="C:\Users\konop\AppData\Local\JetBrains\Rider2020.2\extResources" t="IncludeRecursive" />
-    <e p="C:\Users\konop\AppData\Local\JetBrains\Rider2020.2\resharper-host\Local\Transient\Rider\v202\SolutionCaches\_Tanks_patterns.2132903909.00" t="ExcludeRecursive" />
-    <e p="C:\Users\konop\Documents\GitHub\Tanks_patterns" t="IncludeRecursive">
+    <e p="C:\Users\1\AppData\Local\JetBrains\Rider2020.2\extResources" t="IncludeRecursive" />
+    <e p="C:\Users\1\AppData\Local\JetBrains\Rider2020.2\resharper-host\Local\Transient\Rider\v202\SolutionCaches\_Tanks_patterns.298994923.00" t="ExcludeRecursive" />
+    <e p="C:\Users\1\Documents\Unity Rep\HW_Terrain\Tanks_patterns" t="IncludeRecursive">
       <e p="Assembly-CSharp.csproj" t="IncludeRecursive" />
       <e p="Assets" t="Include">
         <e p="Code" t="Include">
@@ -19,7 +19,7 @@
             <e p="FirePoint.cs" t="Include" />
             <e p="SpawnPoint.cs" t="Include" />
           </e>
-          <e p="TankClass" t="Include">
+          <e p="TankUnit" t="Include">
             <e p="Code" t="Include">
               <e p="TankClass.cs" t="Include" />
               <e p="TankCOntroller.cs" t="Include" />
