+++ .idea/.idea.Tanks_patterns/.idea/workspace.xml
@@ -44,12 +44,22 @@
         <option value="$PROJECT_DIR$/Assets/Code/TankClass/Code/TankView.cs" />
         <option value="$PROJECT_DIR$/Assets/Code/TankClass/Interfaces/ITankFire.cs" />
         <option value="$PROJECT_DIR$/Assets/Code/TankClass/Interfaces/ITankFactory.cs" />
-        <option value="$PROJECT_DIR$/Assets/Code/EventHandlers/FireEvent.cs" />
-        <option value="$PROJECT_DIR$/Assets/Code/GameStarter.cs" />
         <option value="$PROJECT_DIR$/Assets/Code/TankClass/Code/TankFire.cs" />
         <option value="$PROJECT_DIR$/Assets/Code/Markers/FirePoint.cs" />
-        <option value="$PROJECT_DIR$/Assets/Code/TankClass/TankFactory.cs" />
         <option value="$PROJECT_DIR$/Assets/Code/TankClass/Code/TankClass.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankClass/TankFactory.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Code/TankFire.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Code/TankModel.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Interfaces/ITakeDamage.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Interfaces/ITankControl.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Interfaces/ITankFactory.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/GameStarter.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/EventHandlers/FireEvent.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/TankFactory.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Code/TankClass.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Interfaces/ITankView.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Code/TankController.cs" />
+        <option value="$PROJECT_DIR$/Assets/Code/TankUnit/Code/TankView.cs" />
       </list>
     </option>
   </component>
@@ -64,7 +74,14 @@
     <property name="Discover.Launch.Via.Unity" value="true" />
     <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
     <property name="WebServerToolWindowFactoryState" value="false" />
+    <property name="node.js.detected.package.eslint" value="true" />
+    <property name="node.js.detected.package.tslint" value="true" />
+    <property name="node.js.path.for.package.eslint" value="project" />
+    <property name="node.js.path.for.package.tslint" value="project" />
+    <property name="node.js.selected.package.eslint" value="(autodetect)" />
+    <property name="node.js.selected.package.tslint" value="(autodetect)" />
     <property name="settings.editor.selected.configurable" value="editor.preferences.smartKeys_rider" />
+    <property name="vue.rearranger.settings.migration" value="true" />
   </component>
   <component name="RunManager" selected="Attach to Unity Editor.Attach to Unity Editor &amp; Play">
     <configuration name="Attach to Unity Editor &amp; Play" type="UNITY_DEBUG_RUN_CONFIGURATION" factoryName="UNITY_ATTACH_AND_PLAY" show_console_on_std_err="false" show_console_on_std_out="false" port="50000" address="localhost">
@@ -98,8 +115,13 @@
       <workItem from="1635266108386" duration="7331000" />
       <workItem from="1635277132784" duration="4194000" />
       <workItem from="1635356093834" duration="4898000" />
+      <workItem from="1635406248232" duration="1000" />
+      <workItem from="1635406341385" duration="3168000" />
     </task>
     <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
   </component>
   <component name="UnityProjectConfiguration" hasMinimizedUI="true" />
   <component name="UnityUnitTestConfiguration" currentTestLauncher="EditMode" />
@@ -171,10 +193,11 @@
       <screen x="0" y="0" width="1366" height="728" />
     </state>
     <state width="1323" height="250" key="GridCell.Tab.3.right/0.0.1366.728@0.0.1366.728" timestamp="1635281375015" />
-    <state x="495" y="49" key="RiderGenerateDialog" timestamp="1635271226465">
-      <screen x="0" y="0" width="1366" height="728" />
+    <state x="928" y="71" key="RiderGenerateDialog" timestamp="1635408587971">
+      <screen x="0" y="0" width="2560" height="1050" />
     </state>
     <state x="495" y="49" key="RiderGenerateDialog/0.0.1366.728@0.0.1366.728" timestamp="1635271226465" />
+    <state x="928" y="71" key="RiderGenerateDialog/0.0.2560.1050@0.0.2560.1050" timestamp="1635408587971" />
     <state x="184" y="9" key="SettingsEditor" timestamp="1635357895264">
       <screen x="0" y="0" width="1366" height="728" />
     </state>
