+++ Assets/Code/TankUnit/Code/TankClass.cs
@@ -1,0 +1,15 @@
+﻿using CommonClasses;
+using Markers;
+using TankUnit.Interfaces;
+using UnityEngine;
+
+namespace TankUnit.Code
+{
+    public abstract class TankClass : MonoBehaviour
+    {
+        public static TankView TankView;
+        public ITankFire TankFire;
+        public ITankControl TankControl;
+       public Health Health { get; set; }
+    }
+}
