+++ Assets/Code/TankUnit/Code/TankController.cs
@@ -1,15 +1,17 @@
-﻿using TankClass.Interfaces;
+﻿using TankUnit.Interfaces;
 
-namespace TankClass.Code
+namespace TankUnit.Code
 {
-    public class TankController : ITankControl
+    public class TankController : ITankControl 
     {
         private readonly ITankModel _tankModel;
-        private ITankView _tankView;
-        public TankController(ITankModel tankModel, ITankView tankView)
+        private readonly ITankView _tankView;
+
+        public TankController(ITankModel tankModel, ITankView tankView, bool ai)
         {
             _tankModel = tankModel;
             _tankView = tankView;
+            _tankView.ai = ai;
         }
         
         public void Move()
