+++ Assets/Code/TankUnit/Interfaces/ITakeDamage.cs
@@ -1,4 +1,4 @@
-namespace TankClass.Interfaces
+namespace TankUnit.Interfaces
 {
     public interface ITakeDamage 
     {
