+++ Assets/Code/TankUnit/Interfaces/ITankFire.cs
@@ -1,6 +1,4 @@
-﻿using UnityEngine;
-
-namespace TankClass.Interfaces
+﻿namespace TankUnit.Interfaces
 {
     public interface ITankFire
     {
